backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
qdrant-client==1.6.9
sentence-transformers==2.7.0
langchain==0.2.16
langchain-community==0.2.16
requests==2.31.0
sqlparse==0.4.4
pydantic==2.5.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
playwright==1.40.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0